@@ -16,7 +16,6 @@ long_description_content_type = text/markdown
 platform = any
 classifiers =
     Programming Language :: Python
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
@@ -41,7 +40,7 @@ install_requires =
     requests>=2.27.1,<2.29.0
     requests-oauthlib~=1.3.1
     ujson>=5.4,<5.6
-python_requires = >=3.7, <3.13
+python_requires = >=3.8, <3.13
 
 [options.extras_require]
 dev =
