@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -44,18 +43,18 @@ source_suffix = {'.rst': 'restructuredtext'}
 master_doc = 'index'
 
 # General information about the project.
-project = u'WikibaseIntegrator'
-copyright = u'%d, LeMyst' % datetime.now().year
-author = u'LeMyst and WikibaseIntegrator contributors'
+project = 'WikibaseIntegrator'
+copyright = f'{datetime.now().year}, LeMyst'
+author = 'LeMyst and WikibaseIntegrator contributors'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = u'0.12.5.dev0'
+version = '0.12.5.dev0'
 # The full version, including alpha/beta/rc tags.
-release = u'0.12.5.dev0'
+release = '0.12.5.dev0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
