@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
+from typing import TYPE_CHECKING, Any
 
 from wikibaseintegrator.models.basemodel import BaseModel
 from wikibaseintegrator.models.snaks import Snak, Snaks
@@ -12,24 +12,24 @@ if TYPE_CHECKING:
 
 class References(BaseModel):
     def __init__(self) -> None:
-        self.references: List[Reference] = []
+        self.references: list[Reference] = []
 
     @property
-    def references(self) -> List[Reference]:
+    def references(self) -> list[Reference]:
         return self.__references
 
     @references.setter
-    def references(self, value: List[Reference]):
+    def references(self, value: list[Reference]):
         self.__references = value
 
-    def get(self, hash: Optional[str] = None) -> Optional[Reference]:
+    def get(self, hash: str | None = None) -> Reference | None:
         for reference in self.references:
             if reference.hash == hash:
                 return reference
         return None
 
     # TODO: implement action_if_exists
-    def add(self, reference: Optional[Union[Reference, Claim]] = None, action_if_exists: ActionIfExists = ActionIfExists.REPLACE_ALL) -> References:
+    def add(self, reference: Reference | Claim | None = None, action_if_exists: ActionIfExists = ActionIfExists.REPLACE_ALL) -> References:
         from wikibaseintegrator.models.claims import Claim
         if isinstance(reference, Claim):
             reference = Reference(snaks=Snaks().add(Snak().from_json(reference.get_json()['mainsnak'])))
@@ -42,19 +42,19 @@ class References(BaseModel):
 
         return self
 
-    def from_json(self, json_data: List[Dict]) -> References:
+    def from_json(self, json_data: list[dict]) -> References:
         for reference in json_data:
             self.add(reference=Reference().from_json(reference))
 
         return self
 
-    def get_json(self) -> List[Dict]:
-        json_data: List[Dict] = []
+    def get_json(self) -> list[dict]:
+        json_data: list[dict] = []
         for reference in self.references:
             json_data.append(reference.get_json())
         return json_data
 
-    def remove(self, reference_to_remove: Union[Claim, Reference]) -> bool:
+    def remove(self, reference_to_remove: Claim | Reference) -> bool:
         from wikibaseintegrator.models.claims import Claim
         if isinstance(reference_to_remove, Claim):
             reference_to_remove = Reference(snaks=Snaks().add(Snak().from_json(reference_to_remove.get_json()['mainsnak'])))
@@ -80,7 +80,7 @@ class References(BaseModel):
 
 
 class Reference(BaseModel):
-    def __init__(self, snaks: Optional[Snaks] = None, snaks_order: Optional[List] = None):
+    def __init__(self, snaks: Snaks | None = None, snaks_order: list | None = None):
         self.hash = None
         self.snaks = snaks or Snaks()
         self.snaks_order = snaks_order or []
@@ -110,7 +110,7 @@ class Reference(BaseModel):
         self.__snaks_order = value
 
     # TODO: implement action_if_exists
-    def add(self, snak: Optional[Union[Snak, Claim]] = None, action_if_exists: ActionIfExists = ActionIfExists.REPLACE_ALL) -> Reference:
+    def add(self, snak: Snak | Claim | None = None, action_if_exists: ActionIfExists = ActionIfExists.REPLACE_ALL) -> Reference:
         from wikibaseintegrator.models.claims import Claim
         if isinstance(snak, Claim):
             snak = Snak().from_json(snak.get_json()['mainsnak'])
@@ -122,15 +122,15 @@ class Reference(BaseModel):
 
         return self
 
-    def from_json(self, json_data: Dict[str, Any]) -> Reference:
+    def from_json(self, json_data: dict[str, Any]) -> Reference:
         self.hash = json_data['hash']
         self.snaks = Snaks().from_json(json_data['snaks'])
         self.snaks_order = json_data['snaks-order']
 
         return self
 
-    def get_json(self) -> Dict[str, Union[Dict, List]]:
-        json_data: Dict[str, Union[Dict, List]] = {
+    def get_json(self) -> dict[str, dict | list]:
+        json_data: dict[str, dict | list] = {
             'snaks': self.snaks.get_json(),
             'snaks-order': self.snaks_order
         }
